+++ libs/gltfio/include/gltfio/FilamentAsset.h
@@ -31,8 +31,6 @@ namespace filament {
 
 namespace gltfio {
 
-struct BufferBinding;
-struct TextureBinding;
 class Animator;
 
 /**
@@ -53,7 +51,7 @@ class Animator;
  * and upload data into vertex buffers and index buffers.
  *
  * \todo Only the default glTF scene is loaded, other glTF scenes are ignored.
- * \todo Cameras, extras, and extensions are ignored.
+ * \todo Cameras are ignored.
  */
 class FilamentAsset {
 public:
@@ -153,11 +151,6 @@ public:
      */
     const void* getSourceAsset() noexcept;
 
-    const BufferBinding* getBufferBindings() const noexcept;   //!< \deprecated please use ResourceLoader
-    size_t getBufferBindingCount() const noexcept;             //!< \deprecated please use ResourceLoader
-    const TextureBinding* getTextureBindings() const noexcept; //!< \deprecated please use ResourceLoader
-    size_t getTextureBindingCount() const noexcept;            //!< \deprecated please use ResourceLoader
-
     /*! \cond PRIVATE */
 protected:
     FilamentAsset() noexcept = default;
@@ -169,61 +162,6 @@ public:
     FilamentAsset& operator=(FilamentAsset const&) = delete;
     FilamentAsset& operator=(FilamentAsset&&) = delete;
     /*! \endcond */
-};
-
-/**
- * \struct BufferBinding FilamentAsset.h gltfio/FilamentAsset.h
- * \brief Read-only structure that tells the resource loader how to load a source blob into a
- * filament::VertexBuffer, filament::IndexBuffer, etc.
- *
- * \warning Clients usually do not need to interact with BufferBinding directly, they can use
- * ResourceLoader instead.
- *
- * Each binding instance corresponds to one of the following:
- *
- * - One call to VertexBuffer::setBufferAt().
- * - One call to IndexBuffer::setBuffer().
- */
-struct BufferBinding {
-    const char* uri;      // unique identifier for the source blob
-    uint32_t totalSize;   // size in bytes of the source blob at the given URI
-    uint8_t bufferIndex;  // only used when the destination is a VertexBuffer
-    uint32_t offset;      // byte count used only for vertex and index buffers
-    uint32_t size;        // byte count used only for vertex and index buffers
-    void** data;          // pointer to the resource data in the source asset (if loaded)
-
-    // Only one of the following two destinations can be non-null.
-    filament::VertexBuffer* vertexBuffer;
-    filament::IndexBuffer* indexBuffer;
-
-    bool convertBytesToShorts;   // the resource loader must convert the buffer from u8 to u16
-    bool generateTrivialIndices; // the resource loader must generate indices like: 0, 1, 2, ...
-    bool generateDummyData;      // the resource loader should generate a sequence of 1.0 values
-    bool generateTangents;       // the resource loader should generate tangents
-    bool sparseAccessor;         // the resource loader should apply a sparse data set
-
-    bool isMorphTarget;
-    uint8_t morphTargetIndex;
-};
-
-/**
- * \struct TextureBinding FilamentAsset.h gltfio/FilamentAsset.h
- * \brief Read-only structure that describes a binding between filament::Texture and
- * filament::MaterialInstance.
- *
- * \warning Clients usually do not need to interact with TextureBinding directly, they can use
- * ResourceLoader instead.
- */
-struct TextureBinding {
-    const char* uri;
-    uint32_t totalSize;
-    const char* mimeType;
-    void** data;
-    size_t offset;
-    filament::MaterialInstance* materialInstance;
-    const char* materialParameter;
-    filament::TextureSampler sampler;
-    bool srgb;
 };
 
 } // namespace gltfio
