+++ libs/gltfio/src/FilamentAsset.cpp
@@ -57,22 +57,6 @@ MaterialInstance* const* FilamentAsset::getMaterialInstances() noexcept {
     return upcast(this)->getMaterialInstances();
 }
 
-size_t FilamentAsset::getBufferBindingCount() const noexcept {
-    return upcast(this)->getBufferBindingCount();
-}
-
-const BufferBinding* FilamentAsset::getBufferBindings() const noexcept {
-    return upcast(this)->getBufferBindings();
-}
-
-size_t FilamentAsset::getTextureBindingCount() const noexcept {
-    return upcast(this)->getTextureBindingCount();
-}
-
-const TextureBinding* FilamentAsset::getTextureBindings() const noexcept {
-    return upcast(this)->getTextureBindings();
-}
-
 size_t FilamentAsset::getResourceUriCount() const noexcept {
     return upcast(this)->getResourceUriCount();
 }
